+++ tensor-dyn/src/tensor_base.rs
@@ -86,7 +86,7 @@ impl<T> TensorInfo<T> for _Tensor<T> {
     }
 
     fn size(&self) -> usize {
-        self.layout.size()
+        self.layout.size() as usize
     }
 
     fn shape(&self) -> &tensor_common::shape::Shape {
@@ -120,7 +120,7 @@ impl<T> TensorInfo<T> for &_Tensor<T> {
     }
 
     fn size(&self) -> usize {
-        self.layout.size()
+        self.layout.size() as usize
     }
 
     fn shape(&self) -> &tensor_common::shape::Shape {
